backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
anthropic==0.40.0
pdfplumber==0.11.4
beautifulsoup4==4.12.3
requests==2.32.3
sqlalchemy==2.0.36
python-dotenv==1.0.1
httpx==0.28.0
aiofiles==24.1.0
python-multipart==0.0.12
playwright==1.49.0
reportlab==4.2.5
psycopg2-binary==2.9.10
gunicorn==23.0.0
